+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Honeycomb OpenTelemetry SDK Changelog
 
 ## v.Next
 
+## 2.2.1
+
 * fix: Fix crash caused by initializing CTTelephonyNetworkInfo on a background thread during MetricKit diagnostics. Initialization now happens safely on the main thread.
 * feat: add `networkStatusTrackingEnabled` option to toggle collecting network status attributes